requirements.txt
## The following requirements were added by pip freeze:
anyio==4.3.0
apache-beam==2.54.0
asttokens==2.4.1
attrs==23.2.0
basemap==1.4.1
basemap-data==1.3.2
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
comm==0.2.1
contourpy==1.2.0
crcmod==1.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
facets-overview==1.1.1
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.19.1
fonttools==4.49.0
google-api-core==2.17.1
google-apitools==0.5.31
google-auth==2.28.1
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.23.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.1
google-cloud-datastore==2.19.0
google-cloud-dlp==3.15.1
google-cloud-language==2.13.1
google-cloud-pubsub==2.19.4
google-cloud-pubsublite==1.9.0
google-cloud-recommendations-ai==0.10.8
google-cloud-resource-manager==1.12.1
google-cloud-spanner==3.42.0
google-cloud-storage==2.14.0
google-cloud-videointelligence==2.13.1
google-cloud-vision==3.7.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
h2==4.1.0
hdfs==2.7.3
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.18.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
Js2Py==0.74
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.24.4
oauth2client==4.1.3
objsize==0.7.0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==1.4.2
Pygments==2.17.2
pyjsparser==2.7.1
pymongo==4.6.2
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
timeloop==1.0.2
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.10
wrapt==1.16.0
zipp==3.17.0
zstandard==0.22.0
